refactor(spectral tools): NumberScale moved from au3 lib to spectrogram module
