+++ src/spectrogram/internal/numberscale.h
